requirements.txt
fastapi
uvicorn
python-jose
redis
mysql-connector-python
python-dotenv
